requirements.txt
# ai
langgraph==0.3.18
langchain-core==0.3.46
langchain-openai==0.3.9
langgraph-checkpoint-sqlite==2.0.6

# gigachat
langchain-gigachat==0.3.8
gigachat==0.1.39

# jupy
jupyterlab==4.3.6

# other
python-dotenv==1.0.1
